PYG-110 PYG-112 fix: arrumando associação de tags no apidados

+++ src/main/java/edu/fatec/Porygon/controller/ApiDadosController.java
@@ -27,7 +27,7 @@ public class ApiDadosController {
     }
 
     @GetMapping("/dados/{id}")
-    public ResponseEntity<String> abrirDados(@PathVariable Long id) {
+    public ResponseEntity<String> abrirDados(@PathVariable Integer id) {
         Optional<ApiDados> apiDadosOptional = apiDadosRepository.findById(id);
 
         if (apiDadosOptional.isPresent()) {
